Wait for keyspace switch before completing USE statement
Motivation:

When a USE statement is issued, the session switches all of its
connections to the new keyspace. Until now, this was done
asynchronously, so it was possible for a subsequent statement to not see
the new keyspace. Even though USE at runtime is an anti-pattern, we
sometimes use it in tests and this behavior is really annoying.

Modifications:

Make DefaultSession.setKeyspace return a future, and wait on it before
we complete the statement (like we already do for schema updates).

Result:

The second statement is now guaranteed to see the new keyspace (if both
execute calls are synchronous and run by the same thread).