Move YB partition metadata into the cluster's Metadata class
Summary:
Move YugaByte partitioning information into the cluster metadata class.
The partition map will be refreshed on schema changes and on changes to the node list (and token map).
It is used for the partition-aware load-balancing policy, and will also be used by the YugaByte fork of the Spark-Cassandra connector to
create Spark partitions that match our internal partitioning.

Test Plan: YB CQL Java Tests

Reviewers: robert

Reviewed By: robert

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3425