ENG-3604. Fix read replicas not going to local tserver.
Summary: The Cassandra driver wasn't correctly routing requests to the local tserver for consistent-prefix reads. To fix this, added the notion of READ_REPLICA replicas to the driver to recognize local replicas. Also writes were going remote due to a faulty check for strong consistency during iteration at the policy level, so expanded the notion of strong consistency to include QUORUM and LOCAL_ONE cassandra consistency levels.

Test Plan: Ran read and write workload in a read-only node in a live + read-only cluster.

Reviewers: mihnea, robert

Reviewed By: robert

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D5176